Split autoconfiguration of websocket in separate class
