+++ README.md
@@ -41,13 +41,13 @@ repositories {
 }
 
 dependencies {
-  compile 'com.graphql-java:graphql-spring-boot-starter:5.0.1'
+  compile 'com.graphql-java:graphql-spring-boot-starter:5.0.3'
   
   // to embed GraphiQL tool
-  compile 'com.graphql-java:graphiql-spring-boot-starter:5.0.1'
+  compile 'com.graphql-java:graphiql-spring-boot-starter:5.0.3'
 
   // to embed Voyager tool
-  compile 'com.graphql-java:voyager-spring-boot-starter:5.0.1'
+  compile 'com.graphql-java:voyager-spring-boot-starter:5.0.3'
 }
 ```
 
@@ -56,21 +56,21 @@ Maven:
 <dependency>
     <groupId>com.graphql-java</groupId>
     <artifactId>graphql-spring-boot-starter</artifactId>
-    <version>5.0.1</version>
+    <version>5.0.3</version>
 </dependency>
 
 <!-- to embed GraphiQL tool -->
 <dependency>
     <groupId>com.graphql-java</groupId>
     <artifactId>graphiql-spring-boot-starter</artifactId>
-    <version>5.0.1</version>
+    <version>5.0.3</version>
 </dependency>
 
 <!-- to embed Voyager tool -->
 <dependency>
     <groupId>com.graphql-java</groupId>
     <artifactId>voyager-spring-boot-starter</artifactId>
-    <version>5.0.1</version>
+    <version>5.0.3</version>
 </dependency>
 ```
