+++ gradle.properties
@@ -17,7 +17,7 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
 
-PROJECT_VERSION = 5.0.2
+PROJECT_VERSION = 5.0.3
 PROJECT_GROUP = com.graphql-java
 PROJECT_NAME = graphql-spring-boot
 PROJECT_DESC = GraphQL Spring Framework Boot
