+++ graphql-spring-boot-autoconfigure/src/test/java/com/oembedler/moon/graphql/boot/TestAutoConfiguration.java
@@ -1,0 +1,13 @@
+package com.oembedler.moon.graphql.boot;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Import;
+
+@SpringBootConfiguration
+@Import(GraphQLServletProperties.class)
+@EnableConfigurationProperties
+public class TestAutoConfiguration {
+
+
+}
