+++ graphql-spring-boot-autoconfigure/src/test/resources/application.yml
@@ -1,0 +1,1 @@
+graphql.servlet.mapping: /test
